flax_vit/requirements.txt
absl-py==1.3.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
chex==0.1.5
click==8.1.3
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
dill==0.3.5.1
dm-tree==0.1.7
docker-pycreds==0.4.0
etils==0.8.0
flatbuffers==22.9.24
flax==0.6.1
fonttools==4.37.4
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
jax==0.3.23
jaxlib==0.3.22
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
msgpack==1.0.4
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.3
packaging==21.3
pathtools==0.1.2
Pillow==9.2.0
promise==2.3
protobuf==3.19.6
psutil==5.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scipy==1.9.2
sentry-sdk==1.10.1
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-addons==0.18.0
tensorflow-datasets==4.7.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.10.0
termcolor==2.0.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.13.4
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
